+++ ReleaseNotes.md
@@ -8,8 +8,8 @@
   * Ignoring methods returning objects that we do not have the rights to execute/access.
 * **New Environment Variables**
   * IBS.DESERIALIZATION.DEPTH.LIMIT : This value sets the maximum depth of the deserialization.
-  * IBS.PLUGINS.PATH : The package path in which IBS should search for the plugins.
   * IBS.DESERIALIZATION.DATE.FORMAT : The format in which the date should be deserialized.
+  * IBS.PLUGINS.PACKAGE : The package path in which the IBS should search for the plugins you write.
 
 ## 2.11.16
 * **New Feature** [#3 Include an Assertion Feature](https://github.com/adobe/bridgeService/issues/3). We have now included the possibility for users to define assertions. This allows you to clarify accepted results for the call you make with the IBS.